+++ mobile/android/app/src/main/java/com/hyperfamily/companion/MainActivity.java
@@ -1,0 +1,5 @@
+package com.hyperfamily.companion;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}